LayoutTests/streams/reference-implementation/readable-stream-expected.txt


FAIL ReadableStream can be constructed with no errors ReadableStream constructor should get an object as argument.
FAIL ReadableStream instances should have the correct list of properties assert_array_equals: should have all the correct methods lengths differ, expected 6 got 5
PASS ReadableStream constructor should throw for non-function start arguments 
PASS ReadableStream constructor can get initial garbage as cancel argument 
PASS ReadableStream constructor can get initial garbage as pull argument 
PASS ReadableStream constructor can get initial garbage as strategy argument 
FAIL ReadableStream start should be able to return a promise read is not implemented
TIMEOUT ReadableStream start should be able to return a promise and reject it Test timed out
FAIL ReadableStream should be able to enqueue different objects. read is not implemented
PASS ReadableStream: if start throws an error, it should be re-thrown 
FAIL ReadableStream: if pull rejects, it should error the stream read is not implemented
FAIL ReadableStream: should only call pull once upon starting the stream assert_equals: pull should be called once start finishes expected 1 but got 0
FAIL ReadableStream: should only call pull once for a forever-empty stream, even after reading read is not implemented
FAIL ReadableStream: should only call pull once on a non-empty stream read from before start fulfills read is not implemented
FAIL ReadableStream: should only call pull twice on a non-empty stream read from after start fulfills assert_equals: pull should be called once start finishes expected 1 but got 0
FAIL ReadableStream: should call pull in reaction to read()ing the last chunk, if not draining assert_equals: pull should have been called once after read expected 1 but got 0
FAIL ReadableStream: should not call pull() in reaction to read()ing the last chunk, if draining assert_equals: pull should have been called once after read expected 1 but got 0
FAIL ReadableStream: should not call pull until the previous pull call's promise fulfills read is not implemented
FAIL ReadableStream: should pull after start, and after every read read is not implemented
FAIL ReadableStream: should not call pull after start if the stream is now closed read is not implemented
FAIL ReadableStream: should call pull after enqueueing from inside pull (with no read requests), if strategy allows assert_equals: pull() should have been called four times expected 4 but got 0
TIMEOUT ReadableStream pull should be able to close a stream. Test timed out
FAIL ReadableStream: enqueue should throw when the stream is readable but draining assert_equals: the first enqueue should return true expected (boolean) true but got (undefined) undefined
FAIL ReadableStream: enqueue should throw when the stream is closed assert_throws: enqueue after close should throw a TypeError function "function () { c.enqueue('a'); }" did not throw
FAIL ReadableStream: enqueue should throw the stored error when the stream is errored assert_throws: enqueue after error should throw that error function "function () { c.enqueue('a'); }" did not throw
FAIL ReadableStream: should call underlying source methods as methods read is not implemented
FAIL ReadableStream strategies: the default strategy should return false for all but the first enqueue call assert_equals: first enqueue should return true expected (boolean) true but got (undefined) undefined
FAIL ReadableStream strategies: the default strategy should continue returning true from enqueue if the chunks are read immediately assert_equals: first enqueue should return true expected (boolean) true but got (undefined) undefined
FAIL ReadableStream integration test: adapting a random push source read is not implemented
FAIL ReadableStream integration test: adapting a sync pull source read is not implemented
FAIL ReadableStream integration test: adapting an async pull source read is not implemented
